+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+* An event having been processed by the engine will no longer prevent event propagation [#23](https://github.com/Roblox/focus-navigation/pull/23)
+* Events propagated through the FocusNavigationService will now have an attribute indicating that the event was processed [#23](https://github.com/Roblox/focus-navigation/pull/23)
+
+
 ## 1.0.0 (6-5-2023)
 
 * Fix `Event:cancel()` typing ([#19](https://github.com/Roblox/focus-navigation/pull/19))